Fix pylint warnings
