+++ tools/accuracy_checker/accuracy_checker/adapters/__init__.py
@@ -210,6 +210,7 @@ __all__ = [
     'CTCGreedyDecoder',
     'CTCBeamSearchDecoderWithLm',
     'FastCTCBeamSearchDecoderWithLm',
+    'DumbDecoder',
 
     'QualityAssessmentAdapter',
